# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smista-router HTTP API",
    "description": "Local-first deterministic model-routing API. See docs/api/http-api.md.",
    "contact": {
      "name": "Christian Visintin",
      "email": "christian.visintin@veeso.dev"
    },
    "license": {
      "name": "Elastic-2.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/api/v1/auth/bootstrap": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Handles `POST /api/v1/auth/bootstrap`.",
        "operationId": "bootstrap",
        "responses": {
          "201": {
            "description": "User created and API key issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/llm/models": {
      "get": {
        "tags": [
          "llm"
        ],
        "summary": "Handles `GET /api/v1/llm/models`.",
        "description": "The credential-extraction middleware installs the [`RequestCredentials`]\nextension only when the caller sent at least one provider key, so the\nextractor is optional: its absence simply means no credentials, which still\nlists keyless local providers.",
        "operationId": "listModels",
        "parameters": [
          {
            "name": "X-Smista-Timeout-Ms",
            "in": "header",
            "description": "Per-request timeout, default 10000, max 60000",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListModelsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/llm/providers": {
      "get": {
        "tags": [
          "llm"
        ],
        "summary": "Handles `GET /api/v1/llm/providers`.",
        "description": "Reads the live provider registry off the router and returns one descriptor\nper available provider. The order is unspecified, as the registry is keyed\nby provider id.",
        "operationId": "listProviders",
        "responses": {
          "200": {
            "description": "Available providers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProvidersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Envelope wrapping the error body under an `error` key.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody",
            "description": "The structured error payload."
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "The structured payload of an [`ApiError`].\n\n`details` carries machine-readable context (such as the offending provider\nand model) and is omitted when there is none. It must never contain secrets.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, for example `missing_provider_credentials`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, actionable description of what went wrong."
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional structured context; never carries secrets."
          }
        }
      },
...
      "BootstrapResponse": {
        "type": "object",
        "description": "Response to `POST /auth/bootstrap`, returned once at user creation.\n\n`api_key` is a secret shown only here; the caller must store it securely.\n`Debug` is implemented by hand to redact `api_key`, so the credential is\nnever leaked through `{:?}` formatting, logs, or traces.",
        "required": [
          "user_id",
          "api_key"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "description": "Identifier of the created user, for example `user:abc123`."
          },
          "api_key": {
            "type": "string",
            "description": "Long-lived smista API key. Secret; shown only in this response."
          }
        }
      },
...
      "ListModelsResponse": {
        "type": "object",
        "description": "Response to `GET /llm/models`.",
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelDescriptor"
            },
            "description": "Available models across providers."
          },
          "unavailable": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnavailableProvider"
            },
            "description": "Providers omitted from `models` because listing their models failed,\neach with the reason it dropped out.\n\nThe router queries every configured provider when listing models. A\nprovider that cannot be listed — most often because its credentials are\nmissing or were rejected — is left out of `models` and reported here, so\na caller can tell an incomplete result from a genuinely empty one and act\non the cause. Empty when every configured provider was listed."
          }
        }
      },
      "ListProvidersResponse": {
        "type": "object",
        "description": "Response to `GET /llm/providers`.",
        "required": [
          "providers"
        ],
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderDescriptor"
            },
            "description": "Available providers."
          }
        }
      },
...
      "ModelAuthRequirement": {
        "oneOf": [
          {
            "type": "string",
            "description": "No authentication is needed, e.g. a local model.",
            "enum": [
              "none"
            ]
          },
          {
            "type": "string",
            "description": "An API key is required.",
            "enum": [
              "api_key"
            ]
          },
          {
            "type": "string",
            "description": "An API key may be supplied but is not required.",
            "enum": [
              "optional_api_key"
            ]
          },
          {
            "type": "object",
            "description": "A provider-specific scheme, named by the contained string.",
            "required": [
              "custom"
            ],
            "properties": {
              "custom": {
                "type": "string",
                "description": "A provider-specific scheme, named by the contained string."
              }
            }
          }
        ],
        "description": "How a model is authenticated.\n\nEach variant serializes to its snake_case name; [`Self::Custom`] carries the\nscheme name as its payload (e.g. `{\"custom\": \"aws-sigv4\"}`)."
      },
      "ModelCapabilities": {
        "type": "object",
        "description": "The capabilities a model exposes.\n\nEvery flag defaults to `false`, so a [`Default`] value advertises no\ncapabilities — a conservative starting point a descriptor builds upon.",
        "properties": {
          "streaming": {
            "type": "boolean",
            "description": "The model can stream its response incrementally."
          },
          "tools": {
            "type": "boolean",
            "description": "The model can call tools."
          },
          "json_output": {
            "type": "boolean",
            "description": "The model can be constrained to emit JSON."
          },
          "system_prompt": {
            "type": "boolean",
            "description": "The model honors a separate system prompt."
          },
          "images": {
            "type": "boolean",
            "description": "The model accepts image inputs."
          },
          "reasoning": {
            "type": "boolean",
            "description": "The model performs explicit reasoning."
          },
          "memory": {
            "type": "boolean",
            "description": "The model can drive the memory tool to record and update memory.\n\nSatisfied by tool calling, or — on providers that offer it, such as\nOpenAI structured outputs — by a constrained-output equivalent. The\nrouter gates memory orchestration on this flag."
          }
        }
      },
      "ModelDescriptor": {
        "type": "object",
        "description": "A complete description of a model offered by a provider.\n\nThis is also the shape returned for each model by `GET /llm/models`.",
        "required": [
          "provider",
          "model",
          "local",
          "auth",
          "capabilities",
          "max_context_tokens",
          "default_parameters"
        ],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/Provider",
            "description": "Provider that offers the model."
          },
          "model": {
            "type": "string",
            "description": "Model name, as defined by the provider."
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-friendly name, if different from `model`."
          },
          "local": {
            "type": "boolean",
            "description": "Whether the model runs locally (no network call)."
          },
          "auth": {
            "$ref": "#/components/schemas/ModelAuthRequirement",
            "description": "How the model is authenticated."
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities",
            "description": "What the model can do."
          },
          "max_context_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum number of context tokens the model accepts.",
            "minimum": 0
          },
          "max_output_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Maximum number of tokens the model emits, if bounded.",
            "minimum": 0
          },
          "input_cost_per_million_tokens": {
            "type": "string",
            "description": "Input price per million tokens, serialized as a string for exact precision.",
            "example": "3.00"
          },
          "output_cost_per_million_tokens": {
            "type": "string",
            "description": "Output price per million tokens, serialized as a string for exact precision.",
            "example": "15.00"
          },
          "default_parameters": {
            "$ref": "#/components/schemas/ModelParameters",
            "description": "Default generation parameters applied when none are supplied."
          },
          "provider_options": {
            "type": "object",
            "description": "Provider-specific options, preserved verbatim."
          }
        }
      },
      "ModelParameters": {
        "allOf": [
          {
            "type": "object",
            "description": "Provider-specific knobs, preserved verbatim."
          },
          {
            "type": "object",
            "properties": {
              "temperature": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "float",
                "description": "Sampling temperature."
              },
              "top_p": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "float",
                "description": "Nucleus sampling probability mass."
              },
              "max_tokens": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "description": "Maximum number of tokens to generate.",
                "minimum": 0
              }
            }
          }
        ],
        "description": "Sampling and generation parameters for a model invocation.\n\nThe typed fields are omitted from serialization when unset. Any keys not\nmatching a typed field are collected into [`Self::extra`] and re-emitted as\nthey were, so provider-specific options pass through untouched."
      },
...
      "Provider": {
        "type": "string",
        "description": "A provider identifier, e.g. `anthropic`, `openai`, or `openai-compat:<name>`."
      },
...
      "ProviderDescriptor": {
        "type": "object",
        "description": "A provider that is available for routing.\n\nThis is the shape returned by `GET /llm/providers`, which lists only the\nproviders that are currently available — those configured with usable\ncredentials (or a base URL, for local providers). A provider that is not\navailable is omitted from the listing entirely, so this descriptor carries\nno availability flag.",
        "required": [
          "id",
          "display_name",
          "local"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Provider",
            "description": "The provider this descriptor refers to."
          },
          "display_name": {
            "type": "string",
            "description": "Human-friendly name shown to users."
          },
          "local": {
            "type": "boolean",
            "description": "Whether the provider serves its models locally, with no request leaving\nthe host or network.\n\nThis is the provider-level source of truth for locality: every model the\nprovider offers inherits it, so a provider can never report itself local\nwhile a model it advertises routes to the cloud. The router uses it as a\nrouting discriminant to keep sensitive work on local models."
          }
        }
      },
      "ProviderErrorCategory": {
        "type": "string",
        "description": "Classification of a normalized [`ProviderError`].\n\nEach variant captures a category of failure that the router and API layer\ncan reason about independently of the provider-specific HTTP status or\nSDK error code that produced it. The categories follow the smista.ai\nspecification (Model interfaces → Error handling).",
        "enum": [
          "authentication",
          "context_length",
          "invalid_configuration",
          "invalid_credentials",
          "invalid_request",
          "missing_credentials",
          "model_not_found",
          "provider_unavailable",
          "rate_limit",
          "storage",
          "timeout",
          "unknown",
          "unsupported_capability"
        ]
      },
...
      "UnavailableProvider": {
        "type": "object",
        "description": "A provider left out of [`ListModelsResponse::models`] because its models\ncould not be listed.",
        "required": [
          "provider",
          "reason"
        ],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/Provider",
            "description": "The provider whose models could not be listed."
          },
          "reason": {
            "$ref": "#/components/schemas/ProviderErrorCategory",
            "description": "The classification of the failure that left the provider out."
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable, redacted detail describing the failure, when the\nprovider reported one. Never contains credentials."
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Smista-Api-Key",
        "description": "User API key, presented only at sign-in"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from POST /api/v1/auth/sign-in"
      }
    }
  },
...
}